+++ src/error.ts
@@ -1,5 +1,3 @@
-import * as util from './util';
-
 interface IErrorMessage {
   [index: number]: string;
 }
@@ -11,6 +9,7 @@ export enum ErrorCode {
   E348 = 348,
   E444 = 444,
   E488 = 488,
+  E492 = 492,
 }
 
 export const ErrorMessage: IErrorMessage = {
@@ -20,6 +19,7 @@ export const ErrorMessage: IErrorMessage = {
   348: 'No string under cursor',
   444: 'Cannot close last window',
   488: 'Trailing characters',
+  492: 'Not an editor command',
 };
 
 export class VimError extends Error {
@@ -46,10 +46,6 @@ export class VimError extends Error {
 
   get message(): string {
     return this._message;
-  }
-
-  display(): void {
-    util.showError(this.toString());
   }
 
   toString(): string {