+++ src/mode/mode.ts
@@ -1,4 +1,5 @@
 import * as vscode from 'vscode';
+import { VimState } from '../state/vimState';
 
 export enum ModeName {
   Normal,
@@ -11,6 +12,7 @@ export enum ModeName {
   EasyMotionMode,
   EasyMotionInputMode,
   SurroundInputMode,
+  Disabled,
 }
 
 export enum VSCodeVimCursorType {
@@ -23,8 +25,10 @@ export enum VSCodeVimCursorType {
 }
 
 export abstract class Mode {
-  private _isActive: boolean;
-  private _name: ModeName;
+  public readonly name: ModeName;
+  public readonly cursorType: VSCodeVimCursorType;
+  public readonly isVisualMode: boolean;
+  private readonly _statusBarText: string;
   private static readonly _cursorMap = new Map([
     [VSCodeVimCursorType.Block, vscode.TextEditorCursorStyle.Block],
     [VSCodeVimCursorType.Line, vscode.TextEditorCursorStyle.Line],
@@ -33,23 +37,23 @@ export abstract class Mode {
     [VSCodeVimCursorType.TextDecoration, vscode.TextEditorCursorStyle.LineThin],
     [VSCodeVimCursorType.Native, vscode.TextEditorCursorStyle.Block],
   ]);
+  private _isActive: boolean;
 
-  public text: string;
-  public cursorType: VSCodeVimCursorType;
-
-  public isVisualMode = false;
-
-  constructor(name: ModeName) {
-    this._name = name;
+  constructor(
+    name: ModeName,
+    statusBarText: string,
+    cursorType: VSCodeVimCursorType,
+    isVisualMode: boolean = false
+  ) {
+    this.name = name;
+    this.cursorType = cursorType;
+    this.isVisualMode = isVisualMode;
+    this._statusBarText = statusBarText;
     this._isActive = false;
   }
 
-  get name(): ModeName {
-    return this._name;
-  }
-
   get friendlyName(): string {
-    return ModeName[this._name];
+    return ModeName[this.name];
   }
 
   get isActive(): boolean {
@@ -58,6 +62,14 @@ export abstract class Mode {
 
   set isActive(val: boolean) {
     this._isActive = val;
+  }
+
+  getStatusBarText(vimState: VimState): string {
+    return this._statusBarText.toLocaleUpperCase();
+  }
+
+  getStatusBarCommandText(vimState: VimState): string {
+    return vimState.recordedState.commandString;
   }
 
   public static translateCursor(cursorType: VSCodeVimCursorType) {