+++ src/mode/modeHandler.ts
@@ -1,4 +1,5 @@
 import * as vscode from 'vscode';
+import * as modes from './modes';
 
 import { CommandLine } from '../cmd_line/commandLine';
 import { Configuration } from '../configuration/configuration';
@@ -29,15 +30,6 @@ import {
   TextTransformations,
 } from './../transformations/transformations';
 import { Mode, ModeName, VSCodeVimCursorType } from './mode';
-import { EasyMotionInputMode, EasyMotionMode } from './modeEasyMotion';
-import { InsertMode } from './modeInsert';
-import { NormalMode } from './modeNormal';
-import { ReplaceMode } from './modeReplace';
-import { SearchInProgressMode } from './modeSearchInProgress';
-import { VisualMode } from './modeVisual';
-import { VisualBlockMode } from './modeVisualBlock';
-import { VisualLineMode } from './modeVisualLine';
-import { SurroundInputMode } from './surroundInputMode';
 
 export class ModeHandler implements vscode.Disposable {
   private _disposables: vscode.Disposable[] = [];
@@ -51,19 +43,19 @@ export class ModeHandler implements vscode.Disposable {
   }
 
   constructor() {
-    this.createRemappers();
-
+    this._remappers = new Remappers();
     this._modes = [
-      new NormalMode(),
-      new InsertMode(),
-      new VisualMode(),
-      new VisualBlockMode(),
-      new VisualLineMode(),
-      new SearchInProgressMode(),
-      new ReplaceMode(),
-      new EasyMotionMode(),
-      new EasyMotionInputMode(),
-      new SurroundInputMode(),
+      new modes.NormalMode(),
+      new modes.InsertMode(),
+      new modes.VisualMode(),
+      new modes.VisualBlockMode(),
+      new modes.VisualLineMode(),
+      new modes.SearchInProgressMode(),
+      new modes.ReplaceMode(),
+      new modes.EasyMotionMode(),
+      new modes.EasyMotionInputMode(),
+      new modes.SurroundInputMode(),
+      new modes.DisabledMode(),
     ];
 
     this.vimState = new VimState(vscode.window.activeTextEditor!, Configuration.startInInsertMode);
@@ -100,13 +92,6 @@ export class ModeHandler implements vscode.Disposable {
 
     this._disposables.push(disposable);
     this._disposables.push(this.vimState);
-  }
-
-  /**
-   * create remappers after a configuration change
-   */
-  createRemappers() {
-    this._remappers = new Remappers();
   }
 
   /**
@@ -282,15 +267,15 @@ export class ModeHandler implements vscode.Disposable {
   async handleKeyEvent(key: string): Promise<Boolean> {
     const now = Number(new Date());
 
-    // Rewrite commands.
-    // The conditions when you trigger a "copy" rather than a ctrl-c are
-    // too sophisticated to be covered by the "when" condition in package.json
+    // Rewrite commands
     if (Configuration.overrideCopy) {
+      // The conditions when you trigger a "copy" rather than a ctrl-c are
+      // too sophisticated to be covered by the "when" condition in package.json
       if (key === '<D-c>') {
         key = '<copy>';
       }
 
-      if (process.platform !== 'darwin' && key === '<C-c>') {
+      if (key === '<C-c>' && process.platform !== 'darwin') {
         if (
           !Configuration.useCtrlKeys ||
           this.vimState.currentMode === ModeName.Visual ||
@@ -301,9 +286,13 @@ export class ModeHandler implements vscode.Disposable {
         }
       }
     }
-    if (key === '<C-d>' && !Configuration.useCtrlKeys) {
+
+    // <C-d> triggers "add selection to next find match" by default,
+    // unless users explicity make <C-d>: true
+    if (key === '<C-d>' && !(Configuration.handleKeys['<C-d>'] === true)) {
       key = '<D-d>';
     }
+
     this.vimState.cursorPositionJustBeforeAnythingHappened = this.vimState.allCursors.map(
       x => x.stop
     );
@@ -862,6 +851,8 @@ export class ModeHandler implements vscode.Disposable {
         case 'deleteRange':
           edit.delete(new vscode.Selection(command.range.start, command.range.stop));
           break;
+        case 'moveCursor':
+          break;
         default:
           console.warn(`Unhandled text transformation type: ${command.type}.`);
           break;
@@ -1356,6 +1347,7 @@ export class ModeHandler implements vscode.Disposable {
       // Update all EasyMotion decorations
       this.vimState.easyMotion.updateDecorations();
     }
+
     this._renderStatusBar();
 
     await vscode.commands.executeCommand(
@@ -1370,35 +1362,6 @@ export class ModeHandler implements vscode.Disposable {
     );
   }
 
-  private _createCurrentCommandText(currentMode: ModeName): string {
-    switch (currentMode) {
-      case ModeName.Insert:
-        return '';
-      case ModeName.SearchInProgressMode:
-        return `Searching for: ${this.vimState.globalState.searchState!.searchString}`;
-      case ModeName.EasyMotionInputMode:
-        const state = this.vimState.easyMotion;
-        if (state) {
-          const searchCharCount = state.searchAction.searchCharCount;
-          const message =
-            searchCharCount > 0
-              ? `Search for ${searchCharCount} character(s): `
-              : 'Search for characters: ';
-          return message + state.searchAction.getSearchString();
-        }
-        break;
-      case ModeName.EasyMotionMode:
-        return `Target key: ${this.vimState.easyMotion.accumulation}`;
-      case ModeName.SurroundInputMode:
-        if (this.vimState.surround && this.vimState.surround.replacement) {
-          return this.vimState.surround.replacement;
-        }
-        break;
-    }
-
-    return `${this.vimState.recordedState.commandString}`;
-  }
-
   private _renderStatusBar(): void {
     // change status bar color based on mode
     if (Configuration.statusBarColorControl) {
@@ -1411,16 +1374,14 @@ export class ModeHandler implements vscode.Disposable {
     let text = [];
 
     if (Configuration.showmodename) {
-      text.push('--');
-      text.push(this.currentMode.text.toUpperCase());
+      text.push(this.currentMode.getStatusBarText(this.vimState));
       if (this.vimState.isMultiCursor) {
-        text.push('MULTI CURSOR');
+        text.push(' MULTI CURSOR ');
       }
-      text.push('--');
     }
 
     if (Configuration.showcmd) {
-      text.push(this._createCurrentCommandText(this.vimState.currentMode));
+      text.push(this.currentMode.getStatusBarCommandText(this.vimState));
     }
 
     if (this.vimState.isRecordingMacro) {
@@ -1428,7 +1389,8 @@ export class ModeHandler implements vscode.Disposable {
       text.push(macroText);
     }
 
-    StatusBar.SetText(text.join(' '), this.currentMode.name, true);
+    let forceUpdate = this.currentMode.name === ModeName.SearchInProgressMode;
+    StatusBar.SetText(text.join(' '), this.currentMode.name, forceUpdate);
   }
 
   async handleMultipleKeyEvents(keys: string[]): Promise<void> {