+++ CHANGELOG.md
@@ -1,7 +1,34 @@
 # Change Log
 
-## [v0.10.7](https://github.com/vscodevim/vim/tree/v0.10.7) (2018-01-03)
+## [v0.10.10](https://github.com/vscodevim/vim/tree/v0.10.10)
 
+[Full Changelog](https://github.com/vscodevim/vim/compare/v0.10.9...v0.10.10)
+
+- fix: add tests for compareKeyPressSequence [\#2289](https://github.com/VSCodeVim/Vim/pull/2289) ([jpoon](https://github.com/jpoon))
+- Fix BaseAction.couldActionApply to work with two-dimensional keys array [\#2288](https://github.com/VSCodeVim/Vim/pull/2288) ([jpotterm](https://github.com/jpotterm))
+- refactor: move modehandlermap to own class [\#2285](https://github.com/VSCodeVim/Vim/pull/2285) ([jpoon](https://github.com/jpoon))
+- fix: status bar not updating following toggle [\#2283](https://github.com/VSCodeVim/Vim/pull/2283) ([jpoon](https://github.com/jpoon))
+- Fix: Warnings when retrieving configurations w/o resource [\#2282](https://github.com/VSCodeVim/Vim/pull/2282) ([jpoon](https://github.com/jpoon))
+- fix: \<C-d\> remapping disabled by default. functionality controlled by "handleKeys" [\#2269](https://github.com/VSCodeVim/Vim/pull/2269) ([Arxzin](https://github.com/Arxzin))
+
+## [v0.10.9](https://github.com/vscodevim/vim/tree/v0.10.9) (2018-01-11)
+[Full Changelog](https://github.com/vscodevim/vim/compare/v0.10.8...v0.10.9)
+
+- feature: "h", "l" keybindings for sidebar [\#2290](https://github.com/VSCodeVim/Vim/pull/2290) ([Nodman](https://github.com/Nodman))
+- fix: no need to change cursor if there is no active editor. closes \#2273 [\#2278](https://github.com/VSCodeVim/Vim/pull/2278) ([jpoon](https://github.com/jpoon))
+- fix: fixes circular dependency between notation and configuration [\#2277](https://github.com/VSCodeVim/Vim/pull/2277) ([jpoon](https://github.com/jpoon))
+- fix: show cmd-line errors in status bar. add new E492 error [\#2272](https://github.com/VSCodeVim/Vim/pull/2272) ([jpoon](https://github.com/jpoon))
+- refactor: normalize keys when loading configuration [\#2268](https://github.com/VSCodeVim/Vim/pull/2268) ([jpoon](https://github.com/jpoon))
+
+## [v0.10.8](https://github.com/vscodevim/vim/tree/v0.10.8) (2018-01-05)
+[Full Changelog](https://github.com/vscodevim/vim/compare/v0.10.7...v0.10.8)
+
+- fix\(2162\): handleKeys was previously only handling negation [\#2267](https://github.com/VSCodeVim/Vim/pull/2267) ([jpoon](https://github.com/jpoon))
+- fix\(2264\): go-to-line [\#2266](https://github.com/VSCodeVim/Vim/pull/2266) ([jpoon](https://github.com/jpoon))
+- fix\(2261\): change status bar text for search-in-progress to be more l… [\#2263](https://github.com/VSCodeVim/Vim/pull/2263) ([jpoon](https://github.com/jpoon))
+- fix\(2261\): fix regression. show search string in status bar [\#2262](https://github.com/VSCodeVim/Vim/pull/2262) ([jpoon](https://github.com/jpoon))
+
+## [v0.10.7](https://github.com/vscodevim/vim/tree/v0.10.7) (2018-01-04)
 [Full Changelog](https://github.com/vscodevim/vim/compare/v0.10.6...v0.10.7)
 
 - Stop Silently Failing [\#2250](https://github.com/VSCodeVim/Vim/pull/2250) ([jpoon](https://github.com/jpoon))
@@ -248,7 +275,7 @@
 
 - Fixes \#1573: Backspace at beginning of file causes subsequent operation to nop [\#1577](https://github.com/VSCodeVim/Vim/pull/1577) ([Chillee](https://github.com/Chillee))
 - Fix logo src so logo displays inside VSCode [\#1572](https://github.com/VSCodeVim/Vim/pull/1572) ([cobbweb](https://github.com/cobbweb))
-- fixes \#1449 [\#1571](https://github.com/VSCodeVim/Vim/pull/1571) ([azngeoffdog](https://github.com/azngeoffdog))
+- fixes \#1449 [\#1571](https://github.com/VSCodeVim/Vim/pull/1571) ([squedd](https://github.com/squedd))
 - fixes \#1252 [\#1569](https://github.com/VSCodeVim/Vim/pull/1569) ([xconverge](https://github.com/xconverge))
 - fixes \#1486 :wqa command [\#1568](https://github.com/VSCodeVim/Vim/pull/1568) ([xconverge](https://github.com/xconverge))
 - fixes \#1357 [\#1567](https://github.com/VSCodeVim/Vim/pull/1567) ([xconverge](https://github.com/xconverge))
@@ -391,7 +418,7 @@
 - More surround fixes [\#1289](https://github.com/VSCodeVim/Vim/pull/1289) ([xconverge](https://github.com/xconverge))
 
 ## [v0.5.3](https://github.com/vscodevim/vim/tree/v0.5.3) (2017-02-12)
-[Full Changelog](https://github.com/vscodevim/vim/compare/v0.5.0...v0.5.3)
+[Full Changelog](https://github.com/vscodevim/vim/compare/v0.5.1...v0.5.3)
 
 - fixes \#1258 [\#1286](https://github.com/VSCodeVim/Vim/pull/1286) ([xconverge](https://github.com/xconverge))
 - avoid using user remapping in test mode [\#1278](https://github.com/VSCodeVim/Vim/pull/1278) ([rufusroflpunch](https://github.com/rufusroflpunch))
@@ -405,11 +432,11 @@
 - Fix autoindent when opening a line above [\#1249](https://github.com/VSCodeVim/Vim/pull/1249) ([inejge](https://github.com/inejge))
 - Fixes README spelling mistake [\#1246](https://github.com/VSCodeVim/Vim/pull/1246) ([eastwood](https://github.com/eastwood))
 
-## [v0.5.0](https://github.com/vscodevim/vim/tree/v0.5.0) (2017-01-23)
-[Full Changelog](https://github.com/vscodevim/vim/compare/v0.5.1...v0.5.0)
-
 ## [v0.5.1](https://github.com/vscodevim/vim/tree/v0.5.1) (2017-01-23)
-[Full Changelog](https://github.com/vscodevim/vim/compare/v0.4.10...v0.5.1)
+[Full Changelog](https://github.com/vscodevim/vim/compare/v0.5.0...v0.5.1)
+
+## [v0.5.0](https://github.com/vscodevim/vim/tree/v0.5.0) (2017-01-23)
+[Full Changelog](https://github.com/vscodevim/vim/compare/v0.4.10...v0.5.0)
 
 - Surround [\#1238](https://github.com/VSCodeVim/Vim/pull/1238) ([johnfn](https://github.com/johnfn))
 - Support "gf" in es6 import statements by adding the file extension [\#1227](https://github.com/VSCodeVim/Vim/pull/1227) ([aminroosta](https://github.com/aminroosta))